Refactor socket logic into separate module
